+++ skfem/assembly/basis/facet_basis.py
@@ -75,7 +75,7 @@ class FacetBasis(AbstractBasis):
 
         # by default use boundary facets
         if facets is None:
-            self.find = np.nonzero(self.mesh.f2t[1] == -1)[0]
+            self.find = np.nonzero(self.mesh.f2t[1] == -1)[0].astype(np.int32)
         else:
             self.find = mesh.normalize_facets(facets)
